refactor(articleparamsform, app, index.tsx): extract App component; rename variables, add gap style
Extract App component from the index.tsx file; replace div with main; remove using clsx in the App
component; Rename isOpen and setIsOpen variables to isMenuOpen and setIsMenuOpen in the
ArticleParamsForm component;               Add gap stule to the form in the ArticleParamsForm
