+++ src/index.tsx
@@ -1,48 +1,12 @@
 import { createRoot } from 'react-dom/client';
-import { StrictMode, CSSProperties, useState } from 'react';
-import clsx from 'clsx';
+import { StrictMode } from 'react';
 
-import { Article } from './components/article/Article';
-import { ArticleParamsForm } from './components/article-params-form';
-import {
-	defaultArticleState,
-	ArticleStateType,
-} from './constants/articleProps';
+import { App } from './components/app';
 
 import './styles/index.scss';
-import styles from './styles/index.module.scss';
 
 const domNode = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(domNode);
-
-const App = () => {
-	const [articleParams, setArticleParams] =
-		useState<ArticleStateType>(defaultArticleState);
-
-	function handleApplyParams(updatedArticleParams: ArticleStateType) {
-		setArticleParams(updatedArticleParams);
-	}
-
-	return (
-		<div
-			className={clsx(styles.main)}
-			style={
-				{
-					'--font-family': articleParams.fontFamilyOption.value,
-					'--font-size': articleParams.fontSizeOption.value,
-					'--font-color': articleParams.fontColor.value,
-					'--container-width': articleParams.contentWidth.value,
-					'--bg-color': articleParams.backgroundColor.value,
-				} as CSSProperties
-			}>
-			<ArticleParamsForm
-				articleParams={articleParams}
-				handleApplyParams={handleApplyParams}
-			/>
-			<Article />
-		</div>
-	);
-};
 
 root.render(
 	<StrictMode>
